agent/requirements.txt
# Dependencies for Strands Agent. AI model access requires appropriate
# IAM permissions and Bedrock model access grants.
strands-agents[otel]>=1.30.0
mcp>=1.23.1
bedrock-agentcore[strands-agents]>=1.0.6
aws-opentelemetry-distro>=0.10.0
awslabs.aws-pricing-mcp-server
requests
html2text
